Rename OptionalPropertyFactory.CodeGenerator -> OptionalProperty
OptionalPropertyFactory is now OptionalProperty.Factory
