+++ src/main/java/org/inferred/freebuilder/processor/Analyser.java
@@ -108,7 +108,7 @@ class Analyser {
       new MultisetProperty.Factory(),
       new ListMultimapProperty.Factory(),
       new SetMultimapPropertyFactory(),
-      new OptionalPropertyFactory(),
+      new OptionalProperty.Factory(),
       new BuildableProperty.Factory(),
       new DefaultProperty.Factory()); // Must be last, as it will always return a CodeGenerator
